removed struct CodeParser, moved function `parse` to module `code`
